Simplify IPASIR_API preprocessor code
Defining IPASIR_SHARED_LIB is no longer required when BUILDING_IPASIR_SHARED_LIB
is defined.

On Windows, GCC and Clang treat __declspec(dllexport) as a synonym
for __attribute__ ((dllexport)), so no special cases are required.

The definitions are compatible with all major compilers. If a compiler
is used that does not support the function attributes used in the
IPASIR header, the compiler error message should be straightforward
enough. IPASIR_API then needs to be set as a compiler flag.