Thread safety: simplify requirements on solvers
This moves synchronization complexity to clients, but is no burden
except for clients that actually use multithreading in callbacks, which
should be very rare. On the plus side, the requirements are now much
easier to understand and less errorprone to implement. Also, with
the old specification, solver implementors potentially needed to worry
about memory ordering issues, even for single-threaded solver instances.